Refactor "get_instance_info" gathering
This change re-factors the "get_instance_info" function so that it's
logging everything to a file instead of stdout. This will help
folks with debugging the gate. The function was logging most of
important data to stdout which is only trapped by the main
playbook run and caused someone to load the entire run output
to get useful information about the environment.

In addition to logging everything to a file the custom fact gathering
commands we were running have been removed in-favor of simply running
the ansible setup module which will provide a lot more data in a
format that is far easier to consume.

Change-Id: Icd5d600f58e59c541bca41a21642c2d270315aaf
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>