Clean tempest dir before cloning.
This ensures that the code is recloned and if a venv is used, it is
unpacked into a clean clone. If a venv is unpacked multiple times (eg
multiple runs of the playbook) into the same dir then that venv may
become inconsistent.

This is a minimal fix for liberty as the full patch didn't backport
cleanly.

Bug: #1592509
Change-Id: I6f3c36a8150b83afabae8d397d5fc7340dbc93fd
Related-to: 13edcc092597b7109d0f75435ab4d07d56149998