Add the galera_client dependency to the repo build
This change is required because without it the client libraries
used to build the MySQL-Python packages will depend on the base
system and they will be wrong.

In the original backport, galera_client was set as a dependency
of the repo_build role. However, due to galera_client's
dependency on pip_lock_down in liberty, the galera_client role
must be invoked with pip_no_index=false.

This is an example of the issue:
  http://paste.openstack.org/show/543424/

This is the same command when this fix is in place:
  http://paste.openstack.org/show/543436/

Backport for change I6536993523f02ba49da741b9ea21535eb33fe402

Change-Id: I8e62bf2e09217e5fb47d021dfa7f754c09297c62
Closes-Bug: #1565128