Add SNI support via OS/python packages
To ensure that the Ansible venv has the right packages installed,
requirements.txt is updated with the appropriate packages.

To ensure that the OS for the deployment host is appropriately
prepared for the Ansible venv creation, to install pip, and
python packages into the venv, OS packages are installed.

Fedora install support is removed from the script as it is an
unsupported platform and unmaintained section.

Change-Id: Ib4e219294f5fd93c6a50814955ab97424a79dafd