Add SNI support via OS packages
To ensure that the OS for the deployment host is appropriately
prepared to install Ansible, to install pip, OS packages are
installed.

Fedora install support is removed from the script as it is an
unsupported platform and unmaintained section.

This is a combined, but selective backport of:
- https://review.openstack.org/346353
- https://review.openstack.org/348860
- https://review.openstack.org/348970

The backport ensures accurate OS detection and appropriate
package installation for Ansible outside of a python venv.

Change-Id: Ib4e219294f5fd93c6a50814955ab97424a79dafd