Remove return_netmask function
The return_netmask function was provided as a way to make the inventory
script compatible with the old inventory schema. Since this
compatibility has been broken for about 2 major releases, it can be
simplified.

The function assumed that either the container dictionary would have the
old netmask, or a new netmask would exist, so those are the only return
options. Were neither defined, None would be returned as a standard
Python behavior. Thus, simply using the 'netmask' variable is a viable
alternative; if it's present, it will be used, if it's None, the
behavior is the same as before.

Change-Id: Ibdd3ca9dedbb006c2a668939a931c9547160d56a