Add haproxy_service_enabled boolean
Allow endpoints to be disabled in the haproxy_default_services
dict.

This is needed specifically for https://review.openstack.org/#/c/340175/
which requires that the nova-metadata-api service move to the
compute nodes when the networking-calico neutron plugin is used.

In such an environment, the metadata nodes serve requests only for
the local hypervisor and LB endpoints for the metadata service
are not needed.

The corresponding flag disabling haproxy_service_enabled will be
added to the Calico change.

Change-Id: I584fe2647ba0d6a70908e55168360752a08261c5