Prevent syslog leakage
To prevent leakage to syslog this commit adds local7 to the list
of excluded facilities in the rsyslog client default configuration
file. This was done because local7 is used as a facility for
openstack services to send log messages and by not excluding it
in the default config file it was sending the log messages to
syslog and the correct log file for the services.

Manual backport of Ia938c1266dac51e6f278cadc8272ac98bb270044.

Partial-Bug: 1588051
Co-Authored-By: Andrew Meserole <andrew.meserole@rackspace.com>
Change-Id: I57f2941ccf2d099cdc96798c7902fde8cc2e24ea