Update manual upgrade docs with new galera steps
The run-upgrade playbook was updated to incorporate better control over
when the galera containers are restarted during an upgrade.
This is the companion commit where the manual upgrade steps are updated
to match the steps in run-upgrade, and also to update the version of
OSA that must be used to perform these new steps.

Change-Id: Iadac8e2d24cdad5f7519d85d35aec0d4022450fb