[DOCS] Fix the appendix order
This change fixes the appendix order in the install-guide-revised-draft.
Makes sure that each appendix has a letter.

This also removes the Nuage and PlumGrid documentation out.  These are
being moved directly to the Neutron role project.  The intention is that
each role project maintains the various configurations that it supports.

Change-Id: Ib80899be923f63e1060a9685b0d396c2a4fcac8b
Closes-Bug: 1611906