Fix rabbitmq issues in run-upgrade.sh script
This commit updates run-upgrade.sh to address the following two issues:

1. Ensure rabbitmq containers are restarted after their hostnames are
   changed, not before.

Currently, we restart rabbitmq containers before the updated hostname
is set in setup-hosts.yml, which results in the rabbitmq-install.yml
playbook failing to upgrade rabbitmq.  This commit moves the rabbitmq
container restarts to after setup.yml is run.

2. Ansible stores the ansible_hostname in its fact cache.
   ansible_hostname is used for the join cluster of rabbitmq.

If the ansible_hostname is changed during a playbook (like during the
upgrade process), the fact cache has to be cleared. That's what this
commit adds: after changing the hostname and restarting the container
with the new hostname, we can flush the fact cache and go ahead: the
next connection will populate the fact cache with the new
ansible_hostname.

NOTE(mattt): As we are running ansible_fact_cleanup.yml 2x, we have to
             create a symlink for each instance it's getting called so
             that the marker file is created with a unique name.  This
             ensures that re-running ansible_fact_cleanup.yml multiple
             times does not result in it getting skipped after the first
             invocation.

Change-Id: Ie97501a87f98ee88ddbfb2734db0642e7e80fae0
Signed-off-by: Jean-Philippe Evrard <jean-philippe.evrard@rackspace.co.uk>