Set use_stderr to False for all services
oslo.log has a setting to log to stderr as well as a log file and
defaults it to True. This causes OpenStack Ansible to log each service's
logs twice: once to /var/log/{{ service }}/{{ binary-name }}.log and
once to /var/log/upstart/{{ binary-name }}.log because upstart views
each log event on stderr as a priority. Defaulting this to False means
we reduce the duplication of log events on each container and that the
upstart logs are actually informative rather than confusing.

Change-Id: I9d2b2ca74fa75b7ec38de4b6fe01fdf73af03fe2
Closes-bug: 1588051