Restrict Ansible fact gathering to base subset
In Ansible 2.x the fact gathering process includes facts gathered
by facter [1] and ohai [2]if they are available. While this may be
useful in some environment, OpenStack-Ansible does not make use of
any of these facts in any of its playbooks/roles and the facts
gathered are therefore just an extra overhead.

This patch removes the facter and aohai gathering subsets. On a
test AIO this reduced the execution time for fact gathering to the
host and all running containers from 1m5s to 6s.

[1] https://docs.puppet.com/facter/
[2] https://docs.chef.io/ohai.html

This patch also re-implements the selective fact gathering in
run-playbooks that was implemented in
I348a4b7dfe70d56a64899246daf65ea834a75d2a but unintentionally
removed in 9be0662c4f79783d4db8736b89bbd38063635067

Change-Id: I9bfee53571f3dc3a0b0a0950242008e815722180