Automatically detect whether to test all integrated roles
This patch implements an automated determination of whether the
systems deployed by the integrated repository should be tested by
tempest based on whether there are hosts in the applicable host
groups.

This reduces the number of variables that need to be set by CI
system which make use of tempest.

Change-Id: I48a965f6b287b396c3d1c8c1d599374e06708e32