Set a long package cache timeout for OpenStack-CI
OpenStack-CI implements a mechanism for rolling forward apt/deb
package sources in a way that generally does not result in cache
expiry in a short space of time.

The default in all OSA roles is currently 5 minutes, resulting
in several cache refreshes during CI execution. These refreshes
are a waste of time.

The OpenStack-CI mirror updates happen 120 minutes. This patch
opts to set the timeout to 60 mins.

Change-Id: If3c8dabafae2ecbe337197b9b7aa56edf295b5cb