Fix wrong version of pip used in bootstrap
The bootstrap-ansible script is failing with a Python backtrace as it
fails to parse the "ndg-httpsclient>=0.4.2;python_version<'3.0'" line in
requirements.txt properly.

Commit c904de2db512 "Isolate Ansible from the deployment host"
introduced the separation of Ansible into a virtualenv. Even though pip
is updated and placed under /usr/local/bin, the system's package is the
version which is copied into the virtualenv (version 1.5.4 in Ubuntu
14.04).

Commit ab887ee13def "Include python requirements to resolve SNI issue
for Ansible venv" introduced a python_version environment marker to the
ndg-httpsclient line in requirements.txt file. This is a feature which
was added in pip version 6.0 and is therefore not supported by the older
version being run in the virtualenv.

This patch uses the system's version of pip which is installed into the
virtualenv, to update itself to the latest version. This way we can
ensure that we are using the correct version of pip which supports
environment markers and other newer features.

Change-Id: I96536b390ab5bde27397d1ff5de6c16332570e65