Add hostname adjustment to liberty>mitaka upgrades
This change introduces an upgrade playbook that will use the CLI to
disable or delete services/agents entries from several OpenStack
services databases which may have host entries which are not RFC 1034 &
1035 compatible. For the services without a command line available, the
delete of the records in the database is implemented (heat services).

The changes have been added to the run-upgrade script and documenation
is included.

This is an alternative implementation to
https://review.openstack.org/#/c/312274/

Implements: blueprint upgrade-liberty-mitaka
Closes-Bug: #1577245

Change-Id: I083e55a5779bc9917df4b85cc71ba31551d5723a
Signed-off-by: Jean-Philippe Evrard <jean-philippe.evrard@rackspace.co.uk>