Add RC source to scripts library
The ansible.cfg file is now gone and ansible commands that use the OSA
inventory need to source the "/usr/local/bin/openstack-ansible.rc" file
or run commands using the "openstack-ansible" command. If the RC file
is not sourced various race conditions can happen and will cause problems
pre/post deployment while using the various scripts. This change simply
adds a command to source the file when the script lib is called.

Change-Id: I2e0ffdab97b0b7d1afa7aa3ccc03c8f8aa461665
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>