Derive the OpenStack service list from the service file
This patch updates the sources-branch-updater to derive the OpenStack
service list from the service file provided on the CLI.

It will only pick up OpenStack git sources and will deliberately ignore
repositories which include a '-' in the name, as those are drivers or
plugins, not services. Drivers/plugins don't contain conf/policy files
which need updating with every SHA bump.

The requirements repo is also deliberately ignored as it contains no
services which require conf/policy file updates.

Change-Id: I51ad9e5e48da18cb46ccf9cfe12118c19c2a1f53