Fix ml2 ports after an upgrade
During the upgrade, new ports will be created, because there are new
agents configured (change of the hostnames to be RFC1034 compliant).

We have the choice to either remove old ports or change the new ports.
Working with the new ports and removing the old ports has an impact on
the instances: they will need to wait for dhcp end of lease to get their
updated dns (because the dns would be on a new port). This cause issues
with name resolution.

This commit takes the other approach to reduce instance dns resolution
downtime, by disabling the new agents and deleting them from the nets
holding a agent with the old hostname (those networks have a port in
BUILD state after an upgrade). The old dhcp port will then be correctly
registered back in neutron.

Change-Id: I1f04a95b94444791854afc003398d2c55f46502a
Signed-off-by: Jean-Philippe Evrard <jean-philippe@evrard.me>