Added the option to copy changes between stock env.d and repo env.d[M->N]
The location of the env.d files for the Newton release has been changed
to the inventory folder and needs to be updated in the existing upgrade
playbook.

This issue is fixed in the current patch. Also the playbook only copies
changes between the stock env.d and files in /etc/openstack_deploy/env.d
located in the root repo.

Change-Id: I9b368617565931306bd075a0f5626466ad924493
Closes-Bug: #1620691