Configure Ceilometer middleware for Gnocchi-Swift
When using Swift storage for Gnocchi we want to add a filter to the
Ceilometer middleware for Swift to exclude reporting traffic caused
by Gnocchi storing data into Swift to avoid feedback-loops. This
is typically done by isolating the Gnocchi in a dedicated Project in
Keystone.

The best place to filter this traffic is in the middleware as
opposed to filtering in the Ceilometer Collector.

Change-Id: I67b5bff394ad35cf95d10ba32d602954799d7348
Depends-On: If20248494d4e6c127db4ffdf77ae43482f64fe58