Gnocchi identities created before Swift playbook
When Swift is used for Gnocchi storage we need to add the project
to which the Gnocchi service user is assigned to an exclude list for
Swift's telemetry middleware. That requires the Gnocchi identities
to be created before Swift playbook execution but Gnocchi service
install depends on an existing Swift service when it uses Swift
for storage. This change ensures that those identities exist before
Swift playbook execution.

Depends-On: If66fa8c3aa44d44019e3cedb5033ab83a65bd6ab
Depends-On: If4fa5c1c4401911057f99488694d74f5d83330a8
Change-Id: I0b6b53d5a4ad7423058c382160cc5009e0b3d866