Fix log path option
The log path option is attempting to log to a directory that may not
exist. To fix this the "openstack-ansible.rc" file will now check if
the log path directory exists and if it is missing an attempt will be
made to create it. Should there be any issue in creating the log
path directory the variable will be unset allowing Ansible to continue
without emitting an error message.

Change-Id: Ic7827f35df3a7005b24db79e5b374f5e4ed1282b
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>