Define networking for Multi-node environments
This commit makes makes possible to have multi-node environments, by
having a different way to configure networking.

By default, if multi-node is enabled, vxlan encapsulation of the
usual traffic will be enabled.
If encapsulation is disabled, but multi-node is enabled, the
encapsulation sections will be skipped. The deployer should therefore
define the variables to plug the bridges into the appropriate interfaces
with the variables ``bootstrap_host_bridge_(mgmt|vxlan|storage)_ports``

Change-Id: I6a88b2afa76130575e67601628808b7a573aa834
Signed-off-by: Jean-Philippe Evrard <jean-philippe.evrard@rackspace.co.uk>