Enabled conversion of existing db  and tables to utf8_general_ci [M->N]
While upgrading from Mitaka to Newton, the db-collation-alter playbook
attempts to convert the existing db and tables to utf8_general_ci but
fails.
The reason for this was because the FOREIGN_KEY_CHECKS variable in
the playbook was being set/reset only for a particular session
(i.e. a play) and once the play got over the variable was again reset
to its default value(i.e. 1).

This patch ensures that the foreign key checks are indeed disabled
throughout the playbook by setting the global variable instead of
the local one.

Change-Id: Ifed6279fcb5f418c1e1a348739bcd50ade42464f
Closes-Bug: #1621983