Properly namespace cinder_storage_address
In the os_cinder role we have a "cinder_storage_address" variable which
is never set, this is meant to be the "storage_address" variable passed
to the role by the playbook.

The "storage_address" variable which is being used was not namespaced
and could cause issues with other storage services that have a similarly
named variable. Additionally, this ensures the "cinder_storage_address"
variable is not set to 127.0.0.1, which can cause confusion, as the
iscsi_ip_address setting is set to this value.

Secondly, we should utilise the appropriate "storage_address" variable
for cinder backends when we reference them in the sample configuration.
This PR adjusts that documentation, although using the original value
will still work as it did before.

Change-Id: I2b5f78f30d98e8bb49968e32f1b6bbc646014abd