Add lxc_host dynamic group to inventory.
Previously, it was difficult to target only the hosts holding LXC
containers. This change will add a host to the implicit 'lxc_hosts'
group any time a container is created in association with that host.

Users cannot define this group in their configuration. Doing so will
result in an runtime error.

This approach allows for specific targeting without using complicated
Jinja calculations as seen in commit
76695012e6f45fe44beec6a5f67abee43881fbc2. As such, the work in that
review has been replaced with this change.

Change-Id: I10faf5880a61abcc77f017573e0fd76d033224eb