Aodh should inherit service_region
All of the Openstack services deployed in the AIO inherit
a global "service_region" var, which can be used to set the
Openstack region for all services globally. Aodh does not currently
inherit this var so the result when building with a specific region
is an install that has all services in the correct region, and Aodh
in RegionOne.

Change-Id: I501c303673a028b496963a9f0d7ab33670aec6d4