Move inventory management code to enable imports
In order to import code from inventory-manage.py, such as for testing
purposes, the file must either be renamed or some very brittle hacks are
necessary to fake an import. This is due to the '-' character in the
file name.

To accomodate both the goal of importable code and maintained
compatibility, the contents of inventory-manage.py have been moved to
manage_inventory.py, and inventory-manage.py simply imports that code
and runs it.

No new functionality is added.

Change-Id: I60e73bed699b159f47e6c5b46d0787cf15fe1a4f