Remove search_regex from mariadb port check
The 'galera-cluster-rolling-restart.yml' upgrade playbook contains a
race condition when checking that the mariadb port is available after
restarting a galera container. The port can be open but not yet responsive
which will lead to a 'Connection reset by peer' socket error, ending the
playbook run.

Remove the search_regex to avoid prematurely communicating over the
socket. The existing subsequent task will validate that the service is
behaving correctly.

Change-Id: I47f7a66ff9acdf52bf1daeb005c0265fcd1bf7bd