[DOCS] Added release-name as a watermark to Docs.
Added a water mark to install-guide, developer-docs, upgrade-guide
and releasenotes to make clear to deployers which release they're
working with.

Modified conf.py that adds watermark to the documentation according
to particular branch. If the current branch is master than
watermark is 'Pre-release' and if the current branch is
stable/<release-name> than watermark is <release-name>.

As global-requirements does not have sphinxmark included in
stable/newton [1], the requirements is added in tox.ini.

I will replicate this patch to all the role documentation once its
gets merged to master.

[1] https://review.openstack.org/376429

Closes-Bug:1624489
Change-Id: If454938abb4990764076436cd428673b456eab3c