Update all SHAs for 14.0.0
This patch updates all the roles to the latest available stable
SHA's, copies the release notes from the updated roles into the
integrated repo, updates all the OpenStack Service SHA's, and
updates the appropriate python requirements pins.

Some adjustments are made to the sources-branch-updater script
for improved error detection and efficiency. Also, based on
https://review.openstack.org/371999 the ternary search/replace
is fixed.

Tempest and Rally repo sources are co-located in the repo_packages
files as neither of them implement stable/* branches.

The nova consoles (novnc/spice) are implemented in their own
repo_packages file in order to simplify doing the SHA updates.

Gnocchi implements a different style of releasing stable branches
so it is also implemented in its own repo_packages file.

Change-Id: I6fa4afab4203c9902ce70fa6bc49ad1dddb560b3