Change default log bind mount to be optional
The default log bind mount will cause service disruption as containers
will need to be rebooted on top of the fact that some services will not
start until log directories are created for the service to write to.
While this is not an issue on greenfield it is a problem on all
upgrades.

Closes-Bug: #1625722
Change-Id: Ic2345012be9d6b46a33da6bd03ccc397d5655a50
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>