Add debug logging to dynamic inventory
Currently debugging the dynamic inventory file can be problematic,
requiring manual intervention with a debugger to step through function
calls and inspect data structures.

This patch adds a logging facility for outputting detailed debug logs.
The output is not intended for end users, and will not be called by
Ansible runs. Rather, it is meant for running manually to aid
contributing developers.

Change-Id: Ie160c556678859abf733ef23727a9f9bff48ff8b