Add support for the Ceph storage driver in Gnocchi
We call ceph_client before calling the os_gnocchi role. This ordering is
important because Gnocchi needs Ceph access to be ready to go when it
runs its DB sync task.

We pass an empty string to `openstack_service_venv_bin`. This is done
because the venv would not exist yet when the ceph_client role runs.
Symlinks for the Ceph Python libraries are created by the os_gnocchi
role itself.

Change-Id: Ia66c53f7ab6bb242b5aab1c5d01b96ca2fd7df97