Add curl to utility distro packages
This patch adds curl to the packages installed into
the utility container to ensure that the task to
get the list of python clients works properly.

The patch also renames the package list variable name
to the new convention and moves the declaration of it
into group_vars so that it is overridable.

Change-Id: If82485f99fe67d2c1957320c61f1231128e5ce50