Remove the redundant exit tasks from the gate
The gate job is currently trying to sync and archive log files twice
this extends runtime and creates broken files with very long repeating
file extensions. The exit tasks are already being trapped within the
gate script and do not need to be in the specific exit failue/success
functions.

Change-Id: Ib7eb9cb609c31f2fb8fa1e02f1cda06e6abd2153
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>