Log the ansible runtime report to a file
The ansible runtime report was going to standard out which makes it log
to console. This report is helpful to know how many times a playbook ran
and how long it took. Sadly because the console log is so big this
information is being lost. This PR makes it so that the report data is
logged to a file.

Change-Id: I1c531755bb7b3d92401fa0f2489157b76186312d
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>