Always force handler execution
When a task fails in a playbook, the default behaviour results
in notified handlers not running. This can result in a
configuration change being executed, but the service not being
restarted.

This patch changes the default behaviour to force the handlers
to run when a playbook fails for any reason except a host being
unreachable.

Change-Id: I8a16bb1a5c04345e962ab01225f3d972781d9ed7
Reference: http://docs.ansible.com/ansible/playbooks_error_handling.html#handlers-and-failure