Move management code to inventory lib.
This change moves the management code into a library file for
consolidated access to the code. Nothing is changed with the interface
to the inventory-manage script.

The lib/__init__.py file is necessary for python to
import the lib directory as a package, which will become more important
as more code is relocated there.

The code was originally moved to playbooks/inventory/lib, but doing so
made Ansible try to run the python files, which they intentionally will
not do.

Change-Id: I2e31927ec48022c7a504096780977319882c6bdf