Ensure that repo_server/repo_build use same user:group
The repo_build and repo_server roles have the same defaults
in terms of values, but not in terms of variable names. It
is essential that both use the same values, so this patch
ensures that they do use the same values and that a deployer
only needs to change one set.

Change-Id: Ic2a09c2104ddfec8aed5eca44461f416e96dcdb3