Check for configured groups in the environment
When defining a configuration, it's possible that the groups defined in
the openstack_user_config.yml file might not match those defined by the
env.d contents. This isn't strictly a problem - in most cases it's
ignored and will simply result in extra data. However, if keys in the
configuration and the environment are mismatched, it could result in
variable data not being mapped correctly in the resultant inventory.

This patchset extends the -c/--check parameter to compare the configured
groups against the environment's groups, and issue a warning if there is
a mismatch, which should help deployers diagnose some configuration
issues.

The warnings were not added to standard run operations due to the amount
of output generated on stdout by the JSON dump.

Change-Id: I7a7e52b31909e963873063e43e28a5bb9a6963fd