Set default keepalived cidr if none is provided
Even if this file is completely ignored because the role using it
is skipped, ansible parses the whole file. Any undefined variables
will cause a failure. Undefined variable was previously a way to
force failure if it was undefined, but now we have to do differently.

In the meantime, this is a quick fix to make sure we don't have
issues with undefined variables for non-multinode haproxy installs.

Change-Id: Ifd00f4eb0a0a3559d7d3bd3c5a4331798a3c96aa
Closes-Bug: 1631920