Fix value for openstack_host_manage_hosts_file
The current resulting value for the group var calculation of
openstack_host_manage_hosts_file is \"True\"\n instead of
True.

This results in the conditional failing in the openstack_hosts
role because the boolean evaluation of this value results in
False.

This patch removes the use of continuation so that the
resulting value is boolean instead of a string.

Change-Id: I9b240137ab013d5b05d1f154250a77253a913329
Closes-Bug: #1631975