Change the common proxy cache manage tasks to be stateful
In certain situation the proxy cache manager may be down, misconfigured,
or otherwise not available. To ensure a consistent environment and give
the deployer a better build experience, especially in the case of an
AIO, the proxy cache common task has been changed to test for cache
availability and add/remove the cache files based on that test.

Change-Id: Ibd0186d67a844ec7b79f423adfb8104d13f72ca5
Closes-Bug: #1632826
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>