Automatically source the rc file for ansible/ansible-playbook
For Newton onwards, when both ansible and ansible-playbook
commands are run the /usr/local/bin/openstack-ansible.rc file
needs to be sourced before executing the commands.

It's very often the case that we get reports that commands have
failed because deployers are trying to use 'ansible-playbook'
instead of 'openstack-ansible'.

This patch ensures that both these commands execute using the
openstack-ansible wrapper to ensure that the rc file is sourced
appropriately and that /etc/openstack_deploy/user*.yml are
included as extra vars under the right circumstances.

The wrapper only implements the sourced config and extra vars if
the ansible/ansible-playbook commands are executed from the git
clone root directory. This protection has been added in order to
ensure that non-OSA execution is not interfered with.

As learned and applied in https://review.openstack.org/385511 the
order of parameters is changed to ensure that the playbook is
before all other parameters.

Change-Id: I6706dcca6046addb57733b53bc1610644bcbb2d1
Closes-Bug: #1633411