Stop all but one RabbitMQ node prior to upgrade
RabbitMQ nodes must all be stopped prior to a major/minor version
upgrade [1]. The role does this by distinguishing between the upgrader
node and the rest in separate stop and start tasks.

Upgrades can fail when more than one member of rabbitmq_all are not
members of the cluster. This is due to a bug fixed for greenfield
deployments by 21c24789019bf1ee85347356fe6c70b0b9045d7c. The same fix
was not applied to upgrades because major/minor upgrades require all
RabbitMQ nodes to be stopped which is incompatible with serialising the
role in isolation.

This change uses a play to stop all but one of the nodes, prior to
running the rabbitmq_server role, and then serialises the running of
the role so that one node is upgraded at a time. This minimises the
downtime as much as possible while allowing the role to be applied to
one node at a time.

This change is related to an equivalent one [2] made to the test
playbook of the role openstack-ansible-rabbitmq_server.

[1] http://www.rabbitmq.com/clustering.html#upgrading
[2] https://review.openstack.org/#/c/390966/

Change-Id: Icca5cb1a96f83063223b6ddbeb02eeb562b0931b
(cherry picked from commit 76639d6a8625c70152515e608b54d183cdfe383b)