Implement upper constraints for Ansible bootstrap
Currently the Ansible bootstrap does not make use of upper
constraints, resulting in an inconsistent set of package
versions for the same tag.

This patch ensures that the upper constraints are applied
to the Ansible bootstrap.

This, much like the pinning we do elsewhere, ensures that
when OpenStack-Ansible is installed from a tag it uses the
exact same package versions as those that were tested and
it produces consistent results.

Combined backport of:
- https://review.openstack.org/388051
- https://review.openstack.org/390501

Change-Id: I61499631fa211ea482c79b2067d09889d9530b40
(cherry picked from commit f2fc0428e6718a77ab901af4cf930496ec212c0f)