Work around bad libvirt-python wheel
The following package must always build from source:

- libvirt-python:
  A pre-built wheel can be missing libvirt capabilities
  from the installed version of libvirt-bin, leading to
  nova-compute failing to start.

Note that this is a re-implementation of
https://review.openstack.org/388027

The re-implementation is necessary as there are no
group_vars scoped to the repo servers. If these vars
are added to group_vars/all then the compute nodes
would also install from source every time, whereas
it only needs to be done in the repo server's wheel
build process.

Closes-Bug: #1625859
Change-Id: I3197f06839c9705574810ab616dff51924435da5