Ensure that upper constraints are always applied
When the upper constraints are not applied, the resulting
build for the same tag may be different from one day to
the next. This results in an inconsistent experience for
deployers.

This patch sets the upper constraints to be used by the
following:
- https://review.openstack.org/388159 (lxc hosts)
- https://review.openstack.org/388072 (pip install)
- https://review.openstack.org/388230 (keystone install)
  This is a pattern to establish across other roles.

As some pip installs happen before the repo server is up
a common set of tasks is added to a few of the plays to
verify if there is a local copy of constraints available.
If not it will then source the upstream constraints based
on the OpenStack requirements SHA that's been pinned.

Change-Id: I7ddd44be30c610609c1429147646dcb1d3e6ca59
(cherry picked from commit 46258cc435a67fdf4e089111984ff7ae6930ffa7)