Remove the old rabbitmq queues
The upgrade process doesn't remove the existing queues with the old
hostnames (with _). To remove the queues, we need to create a user
with admin rights for every vhosts (the existing users per service
 can't delete queues), and then use the API to delete the queues.

The API is not really user friendly, so we are using a python script
instead. This script is delivered with the management plugin of
rabbitmq, so we just have to fetch it and use it.

Finally, the temporary user is deleted, and its password is
cleaned up.

The removal of the queues will also remove them from the bindings.

Change-Id: Ic67d96fb3dc0ada7bff4d6597ca76977f5228f77
Closes-Bug: #1632098