Resolve random node selection in repo build
In some cases the multiple archetecture support in a multi-node cloud
will select nodes incapable of sync'ing data outside of themselves as
primary build targets. This patch ensures that a known group of
"repo_nodes" is created at run-time and consists of unique hosts with
unique archetectures provided the hostname does not match the first
node within the "repo_all" group.

Closes-Bug: #1633784
Change-Id: Ifdfabf3b2ef147d287263f2c79c676267fbdea3c
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>