Add command to remove IPs from inventory
This change adds support for removing container IP information from
inventory. This might be useful in situations where an environment is
deployed, but network information has changed, or there was a mistake in
subnetting (such as an expanded used_ips range).

Updating configs, such as LXC interface files and HAProxy or other load
balancers is not handled via this command.

Change-Id: Ibad9d587e14c41ecd486ed8668f9f815e348a94a