Fix Ansible bootstrap upper constraint use
When bootstrapping the Ansible runtime venv the version of
pip in the initial venv might be old and unable to understand
the --constraint option.

This patch ensures that the pip, setuptools and wheel upgrade
does not try to use the --constraint option (they're fixed pins
anyway so it's not needed) and then applies the --constraint
option once we know we're at the right version of pip.

Change-Id: Ic482493c800362b896bfe5083b253f1d33d809b5