Added lxd password to the secrets file
In order to deplpoy LXD compute types the lxd password is required. This
change adds the `lxd_trust_password` variable to the default passwords
file so that a strong password is generated by default even if it's not
to be used intitally.

Change-Id: Ifb7ec681043b02bbbbcd53bcba01653097a55253
Signed-off-by: Kevin Carter <kevin.carter@rackspace.com>
(cherry picked from commit bd3d9335279f2737aa4444177d5bf629bb2d3322)