Refactor inventory management filesystem
Segmenting the storage-related operations from the rest of the
inventory management tooling, in support of better maintainability
testability and future support for alternate storage options.

Some changes were made in the course of the refactor to facilitate
future changes.

The include_cwd parameter was removed in the course of the refactor
as it was determined to be unused and unneededduring patch set
review.

Change-Id: Iaa9761a73a626c4f02c7662b119881a79397d690