Move IP logic into separate module
This patch moves the IP logic into a module separate from the overall
code. Ideally, this will serve as a way to abstract IP information
handling to be populated by pluggable backends in addition to existing
logic.

Further work is needed to remove various code smells from tests and provide a
more consistent public API that could be adopted as more generic.

Change-Id: I09a57c43064eca46f776e941cbedb2296f43fbbc