Fix BaseException.message deprecation messages
When running inventory tests locally on Python 2.7.12, the following
warning was displayed for some tests:

/home/nrb/projects/openstack-ansible/tests/test_inventory.py:565:
DeprecationWarning: BaseException.message has been deprecated as of
Python 2.6

This change updates references to exception.message to use the str
method of access, which is the replacement.

A __str__ method was also added the LxcHostsDefined exception to be more
consistent.

Change-Id: I75fd2da4a31b75916f10502134913c8bacbc34f0